requirements.txt
bottle
gevent
gevent-websocket
numpy
pyepics